Add a test for reducer(RED)
