+++ src/reducer.js
@@ -1,0 +1,11 @@
+const initialState = {
+  newId: 100,
+  nameText: '',
+  categoryText: '',
+  addressText: '',
+  restaurants: [],
+};
+
+export default function reducer(state, action) {
+  return state;
+}
